Split form components module into a package
