+++ functional_tests/pages/components/forms/generic.py
@@ -1,0 +1,39 @@
+from selenium.webdriver.common.by import By
+
+from ..base import BaseComponent
+
+
+class FormComponent(BaseComponent):
+    SUBMIT_BUTTON = (By.CSS_SELECTOR, "button[type='submit']")
+
+    @property
+    def _submit_button(self):
+        return self.browser.find_element(*self.SUBMIT_BUTTON)
+
+    @property
+    def submit_button_label(self):
+        return self._submit_button.text
+
+    def submit(self):
+        self._submit_button.click()
+        return self
+
+
+class SearchForm(FormComponent):
+    SEARCH_INPUT = (By.CSS_SELECTOR, "input[type='search']")
+
+    @property
+    def _search_input(self):
+        return self.browser.find_element(*self.SEARCH_INPUT)
+
+    @property
+    def search_input_placeholder(self):
+        return self._search_input.get_attribute("placeholder")
+
+    @property
+    def search_input_aria_label(self):
+        return self._search_input.get_attribute("aria-label")
+
+    def search(self, search_term):
+        self._search_input.send_keys(search_term)
+        return self.submit()
